Move snapshots.go to snapshot
